requirements.txt
asgiref==3.8.1
astroid==3.2.4
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
Django==5.0.7
django-ckeditor-5==0.2.13
django-filter==24.3
django-froala-editor==4.2.1
django-guardian==2.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.8
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
mysqlclient==2.2.4
pillow==10.4.0
platformdirs==4.2.2
pylint==3.2.6
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.0
tzdata==2024.1
urllib3==2.2.2
